+++ apps/commerce-api/src/test/kotlin/com/loopers/domain/user/UserServiceTest.kt
@@ -66,179 +66,6 @@ class UserServiceTest @Autowired constructor(
     }
 
     @Test
-    fun `createUser() throws CoreException when userId already exists`() {
-        // Arrange
-        val userId = "duplicateUser"
-        userService.createUser(userId, "password123!", "User1", LocalDate.now(),
-            "duplicateUser@example.com")
-
-        // Act
-        val exception = assertThrows<CoreException> {
-            userService.createUser(userId, "password123!", "User2", LocalDate.now(),
-                "duplicateUser@example.com")
-        }
-
-        // Assert
-        assertThat(exception.errorType).isEqualTo(ErrorType.CONFLICT)
-    }
-
-    @Test
-    fun `createUser() throws CoreException when password is too short`() {
-        // Arrange
-        val shortPassword = "pass123" // 7 chars (< 8)
-
-        // Act
-        val exception = assertThrows<CoreException> {
-            userService.createUser("testUser", shortPassword, "홍길동", LocalDate.now(),
-                "test@example.com")
-        }
-
-        // Assert
-        assertThat(exception.errorType).isEqualTo(ErrorType.BAD_REQUEST)
-        assertThat(exception.message).contains("8~16자")
-    }
-
-    @Test
-    fun `createUser() throws CoreException when password is too long`() {
-        // Arrange
-        val shortPassword = "password123456789!" // 18 chars (> 16)
-
-        // Act
-        val exception = assertThrows<CoreException> {
-            userService.createUser("testUser", shortPassword, "홍길동", LocalDate.now(),
-                "test@example.com")
-        }
-
-        // Assert
-        assertThat(exception.errorType).isEqualTo(ErrorType.BAD_REQUEST)
-        assertThat(exception.message).contains("8~16자")
-    }
-
-    @Test
-    fun `createUser() accept with exact 8 characters`() {
-        // Arrange
-        val shortPassword = "pass1234" // 8 chars
-
-        // Act
-        val user = userService.createUser(
-            "testUser", shortPassword, "홍길동", LocalDate.now(),
-            "test@example.com"
-        )
-
-        // Assert
-        assertThat(user).isNotNull
-    }
-
-    @Test
-    fun `createUser() accept with exact 16 characters`() {
-        // Arrange                                                                         
-        val password = "password12345678" // 16 chars                                      
-
-        // Act                                                                             
-        val user = userService.createUser(
-            "testUser", password, "testName", LocalDate.now(),
-            "test@example.com"
-        )
-
-        // Assert                                                                          
-        assertThat(user).isNotNull
-    }
-
-    @Test
-    fun `createUser() throws CoreException when password contains birthDate in yyyyMMdd format`() {
-        // Arrange                                                                         
-        val birthDate = LocalDate.of(1990, 1, 1)
-        val password = "pass19900101" // contains 19900101                                 
-
-        // Act                                                                             
-        val exception = assertThrows<CoreException> {
-            userService.createUser("testUser", password, "testName", birthDate,
-                "test@example.com")
-        }
-
-        // Assert                                                                          
-        assertThat(exception.errorType).isEqualTo(ErrorType.BAD_REQUEST)
-        assertThat(exception.message).contains("생년월일")
-    }
-
-    @Test
-    fun `createUser() throws CoreException when password contains birthDate in yyMMdd format`() {
-        // Arrange                                                                         
-        val birthDate = LocalDate.of(1990, 1, 1)
-        val password = "pass900101abc" // contains 900101                                  
-
-        // Act                                                                             
-        val exception = assertThrows<CoreException> {
-            userService.createUser("testUser", password, "testName", birthDate,
-                "test@example.com")
-        }
-
-        // Assert                                                                          
-        assertThat(exception.errorType).isEqualTo(ErrorType.BAD_REQUEST)
-        assertThat(exception.message).contains("생년월일")
-    }
-
-    @Test
-    fun `createUser() throws CoreException when password contains birthDate in MMdd format`() {
-        // Arrange                                                                         
-        val birthDate = LocalDate.of(1990, 1, 1)
-        val password = "password0101" // contains 0101                                     
-
-        // Act                                                                             
-        val exception = assertThrows<CoreException> {
-            userService.createUser("testUser", password, "testName", birthDate,
-                "test@example.com")
-        }
-
-        // Assert                                                                          
-        assertThat(exception.errorType).isEqualTo(ErrorType.BAD_REQUEST)
-        assertThat(exception.message).contains("생년월일")
-    }
-
-    @Test
-    fun `createUser() should accept password that does not contain birthDate`() {
-        // Arrange                                                                         
-        val birthDate = LocalDate.of(1990, 1, 1)
-        val password = "safePassword!" // does not contain any birthDate format            
-
-        // Act                                                                             
-        val user = userService.createUser("testUser", password, "testName", birthDate,
-            "test@example.com")
-
-        // Assert                                                                          
-        assertThat(user).isNotNull
-    }
-
-    @Test
-    fun `createUser() throws CoreException when email format is invalid`() {
-        // Arrange                                                                         
-        val invalidEmail = "invalid-email-format" // no @                                  
-
-        // Act                                                                             
-        val exception = assertThrows<CoreException> {
-            userService.createUser("testUser", "password123!", "testName", LocalDate.now(),
-                invalidEmail)
-        }
-
-        // Assert                                                                          
-        assertThat(exception.errorType).isEqualTo(ErrorType.BAD_REQUEST)
-        assertThat(exception.message).contains("이메일")
-    }
-
-    @Test
-    fun `createUser() should accept valid email format`() {
-        // Arrange                                                                         
-        val validEmail = "valid.email@example.com"
-
-        // Act                                                                             
-        val user = userService.createUser("testUser", "password123!", "testName",
-            LocalDate.now(), validEmail)
-
-        // Assert                                                                          
-        assertThat(user.email).isEqualTo(validEmail)
-    }
-
-    @Test
     fun `getUserByUserId() returns User`() {
         // Arrange
         val userId = "testId"
@@ -254,20 +81,6 @@ class UserServiceTest @Autowired constructor(
     }
 
     @Test
-    fun `getUserByUserId() throws CoreException when User is not found`() {
-        // Arrange
-
-        // Act
-
-        // Assert
-        assertThrows<CoreException> {
-            userService.getUserByUserId("nonExistentUser")
-        }.also {
-            assertThat(it.errorType).isEqualTo(ErrorType.NOT_FOUND)
-        }
-    }
-
-    @Test
     fun `authenticate() returns User with correct password`() {
         // Arrange
         val userId = "testId"
@@ -280,36 +93,5 @@ class UserServiceTest @Autowired constructor(
         // Assert
         assertThat(authenticatedUser).isNotNull
         assertThat(authenticatedUser.userId).isEqualTo(userId)
-    }
-
-    @Test
-    fun `authenticate() throws exception with wrong password`() {
-        // Arrange
-        val userId = "testId"
-        val password = "testPassword"
-        userService.createUser(userId, password, "testName", LocalDate.now(), "test@email.com")
-
-        // Act
-
-        // Assert
-        assertThrows<CoreException> {
-            userService.authenticate(userId, "wrongPassword")
-        }.also {
-            assertThat(it.errorType).isEqualTo(ErrorType.UNAUTHORIZED)
-        }
-    }
-
-    @Test
-    fun `authenticate() throws exception with non-existing User`() {
-        // Arrange
-
-        // Act
-
-        // Assert
-        assertThrows<CoreException> {
-            userService.authenticate("nonExistUser", "wrongPassword")
-        }.also {
-            assertThat(it.errorType).isEqualTo(ErrorType.UNAUTHORIZED)
-        }
     }
 }
